+++ src/ChainDatabase.cpp
@@ -10,7 +10,7 @@ constexpr std::string_view GENESIS_BLOCK = "HenryCoin Genesis";
 
 void write_block(std::ostream& stream, const Block& block)
 {
-    ash::db::write_data<std::uint32_t>(stream, block.index());
+    ash::db::write_data<std::uint64_t>(stream, block.index());
     ash::db::write_data<std::uint32_t>(stream, block.nonce());
     ash::db::write_data<std::uint32_t>(stream, block.difficulty());
     ash::db::write_data(stream, block.data());
